% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{aggregate_results}
\alias{aggregate_results}
\title{Aggregate function}
\usage{
aggregate_results(labels, samples, Results, drop_rows = TRUE, sp_max = NULL)
}
\arguments{
\item{labels}{labels}

\item{samples}{samples}

\item{Results}{Results}
}
\description{
Aggregate function
}
