% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortMat.R
\name{array2obsDF}
\alias{array2obsDF}
\title{Transform Arrays to Observation Data Table}
\usage{
array2obsDF(obs_array)
}
\arguments{
\item{obs_array}{A list containing three arrays: species, dbh, and trees.}
}
\value{
A data.frame with columns siteID, patchID, cohortID, species, dbh, and trees.
}
\description{
This function transforms arrays of species, dbh, and trees back into an observation data table.
}
\examples{
obs_array <- list(
  species = array(c("A", "B"), dim = c(2, 2, 2)),
  dbh = array(c(10, 20, 30, 40), dim = c(2, 2, 2)),
  trees = array(c(100, 200, 150, 250), dim = c(2, 2, 2)))
result <- array2obsDF(obs_array)

}
