% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINN.R
\name{BA_stand}
\alias{BA_stand}
\title{BA_stemsal aera multiplied with number of species}
\usage{
BA_stand(dbh, trees, patch_size_ha = 0.1)
}
\arguments{
\item{dbh}{dbh}

\item{trees}{number of Trees}
}
\description{
BA_stemsal aera multiplied with number of species
}
