% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINN.R
\name{build_NN}
\alias{build_NN}
\title{Build a neural network}
\usage{
build_NN(
  self,
  input_shape,
  output_shape,
  hidden,
  bias,
  activation,
  dropout,
  last_activation = "sigmoid"
)
}
\arguments{
\item{input_shape}{integer Number of predictors.}

\item{output_shape}{integer Number of species.}

\item{hidden}{vector List of hidden layers.}

\item{bias}{vector Boolean values indicating whether to use bias in hidden layers.}

\item{activation}{vector List of activation functions.}

\item{dropout}{float Dropout rate.}

\item{last_activation}{character Last activation function.}
}
\value{
torch.nn.modules.container.Sequential Sequential neural network object.
}
\description{
This function builds a neural network with specified input shape, output shape, hidden layers, bias, activation functions, dropout rate, and last activation function.
}
\examples{
build_NN(input_shape = 2, output_shape = 3, hidden = c(1, 3, 2), bias = TRUE, activation = "relu", dropout = 1, last_activation = "sigmoid")

}
