% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINN.R
\name{coef.FINN}
\alias{coef.FINN}
\title{Returns list of parameters}
\usage{
\method{coef}{FINN}(object, ...)
}
\arguments{
\item{object}{a model created by \code{\link{FINN}}}

\item{...}{nothing implemented yet}
}
\value{
list of weights
}
\description{
Returns list of parameters
}
