% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINN.R
\name{cohort_df2arrays}
\alias{cohort_df2arrays}
\title{Make FINN arrays from observation data frame}
\usage{
cohort_df2arrays(cohort_df, ...)
}
\arguments{
\item{cohort_df}{a data frame with the columns Species, dbh, nTree. Optionally cohortID an be provided.}

\item{...}{nothing implemented yet}
}
\value{
list of cohort arrays dbh, nTree, Species from data frame with dendrometric data
}
\description{
Make FINN arrays from observation data frame
}
