% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortMat.R
\name{CohortMat}
\alias{CohortMat}
\title{Cohort Matrix Class}
\usage{
CohortMat(
  obs_df = NULL,
  dbh = NULL,
  trees = NULL,
  species = NULL,
  dims = c(50, 30, 10),
  sp = 10,
  device = "cpu"
)
}
\arguments{
\item{obs_df}{A data frame containing columns "siteID", "patchID", "species", "dbh", and "trees". If provided, it will be used to initialize the tensors.}

\item{dbh}{A tensor or array representing the diameter at breast height. Defaults to \code{self$dbh}.}

\item{trees}{A tensor or array representing the number of trees. Defaults to \code{self$trees}.}

\item{species}{A tensor or array representing the species. Defaults to \code{self$species}.}

\item{dims}{A numeric vector representing the dimensions of the arrays (sites, patches, cohorts). Defaults to \code{self$dims}.}

\item{sp}{An integer representing the number of species. Defaults to \code{self$sp}.}

\item{device}{A character string specifying the device to use ('cpu' or 'cuda'). Defaults to \code{self$device}.
@description

Convert the tensors to a data frame format}
}
\value{
data.table object
}
\description{
Initialize the CohortMat class
}
\details{
An R6 class for managing cohorts of trees in forest models. This class allows for the initialization, transformation, and manipulation of cohorts represented by arrays of dbh, trees, and species.
}
\section{Fields}{

\describe{
\item{\code{dbh}}{A tensor representing the diameter at breast height for each cohort.}

\item{\code{trees}}{A tensor representing the number of trees in each cohort.}

\item{\code{species}}{A tensor representing the species of each cohort.}

\item{\code{dims}}{A vector representing the dimensions of the arrays (sites, patches, cohorts).}

\item{\code{sp}}{An integer representing the number of species.}

\item{\code{device}}{A character string specifying the device to use ('cpu' or 'cuda').}

\item{\code{dbh_r}}{A numeric array representing the diameter at breast height in R array format.}

\item{\code{trees_r}}{A numeric array representing the number of trees in R array format.}

\item{\code{species_r}}{An integer array representing the species in R array format.}

\item{\code{device_r}}{A character string specifying the device in R format ('cpu' or 'cuda').}

\item{\code{obsDF2arrays}}{Transforms data.table into array}
}}

