% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createInputs.R
\name{createInputs}
\alias{createInputs}
\title{Create inputs for the forest model
This function prepares input data for FINN by processing site and tree data. It resolves siteIDs into array dimensions needed for the FINN simulation.}
\usage{
createInputs(
  site_dt,
  tree_dt,
  Nspecies,
  patchsize = patchsize,
  dbh_binsize = 0.1
)
}
\arguments{
\item{site_dt}{A data.table containing site-level data.}

\item{tree_dt}{A data.table containing tree-level data.}

\item{Nspecies}{An integer specifying the maximum number of species to consider.}

\item{patchsize}{A numeric value specifying the patch size for the simulations. Default is 0.06.}

\item{dbh_binsize}{A numeric value specifying the bin size for diameter at breast height (DBH) when creating initial cohorts. Default is 0.5.}
}
\value{
A list containing the following elements:
\item{env_dt}{A data.table with environmental data.}
\item{obs_dt}{A data.table with observation data.}
\item{initCohort}{An object representing the initial cohorts of trees.}
\item{patchsize}{The patch size used for the simulations.}
\item{Nspecies}{The maximum number of species considered.}
}
\description{
Create inputs for the forest model
This function prepares input data for FINN by processing site and tree data. It resolves siteIDs into array dimensions needed for the FINN simulation.
}
