% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createInputs.R
\name{dbh2ba}
\alias{dbh2ba}
\title{Convert DBH to basal area}
\usage{
dbh2ba(dbh)
}
\arguments{
\item{dbh}{numeric vector of diameters (cm).}
}
\value{
numeric vector of basal areas (m^2).
}
\description{
Computes basal area in m^2 from diameter at breast height in cm:
\deqn{BA = \pi \left(\frac{\mathrm{DBH}}{200}\right)^2}
}
