% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processes_utils.R
\name{extract_env}
\alias{extract_env}
\title{Extract Environmental Data for a Process}
\usage{
extract_env(formula, env)
}
\arguments{
\item{formula}{formula object \code{createProcess}.}

\item{env}{A data frame containing environmental data, which must include columns "siteID" and "year".}
}
\value{
An array of environmental data formatted according to the process object.
}
\description{
This function extracts and transforms environmental data according to a specified process object. The environmental data is formatted into an array suitable for input into the simulation model.
}
\examples{
\dontrun{
env_array <- extract_env(growth_process, env_data)
}
}
