% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{FINN.seed}
\alias{FINN.seed}
\title{Set Seed for Reproducibility in R and Torch}
\usage{
FINN.seed(seed)
}
\arguments{
\item{seed}{An integer value to set as the seed for both R and Torch. This ensures that random operations
in both environments produce consistent results.}
}
\value{
This function does not return a value. It sets the seed internally for both R and Torch.
}
\description{
This function sets the seed for both R's random number generator and Torch's random number generator,
ensuring reproducibility across operations that involve both R and Torch.
}
\details{
The function calls \code{set.seed()} to set the seed for R's random number generator and
\code{torch::torch_manual_seed()} to set the seed for Torch's random number generator. This is useful
for ensuring reproducibility in scripts that rely on both R and Torch for random operations.
}
\examples{
\dontshow{if (torch::torch_is_installed()) withAutoprint(\{ # examplesIf}
FINN.seed(123)
# Now both R and Torch are seeded with 123, ensuring reproducible results
\dontshow{\}) # examplesIf}
}
