% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{groupby_mean}
\alias{groupby_mean}
\title{group by mean}
\usage{
groupby_mean(values, labels)
}
\arguments{
\item{values}{list of value tensors}

\item{labels}{labels}
}
\value{
A \code{torch} tensor of the mean of \code{values} within each unique group in
\code{labels}.
}
\description{
group by mean
}
\keyword{internal}
