% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINN.R
\name{height}
\alias{height}
\title{Calculate the height of a tree BA_stemsed on its diameter at breast height and a global parameter.}
\usage{
height(dbh, parHeight)
}
\arguments{
\item{dbh}{A numeric value representing the diameter at breast height of the tree.}

\item{parHeight}{A numeric value representing the global parameter used in the height calculation.}
}
\value{
A numeric value representing the calculated height of the tree.
}
\description{
This function calculates the height of a tree BA_stemsed on the diameter at breast height (dbh) and a parameter.
}
\examples{
height(30, 0.5)

}
