% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{index_species}
\alias{index_species}
\title{Index species}
\usage{
index_species(pred, species)
}
\arguments{
\item{pred}{predictions}

\item{species}{species index vector, must be int64}
}
\value{
A \code{torch} tensor of \code{pred} gathered along its species dimension by
\code{species}.
}
\description{
Index species
}
\keyword{internal}
