% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINN.R
\name{init_FINN}
\alias{init_FINN}
\title{Initialize the FINN model with specified parameters.}
\usage{
init_FINN(
  sp = self$sp,
  env = self$env,
  device = self$device,
  parHeight = self$parHeight,
  parGrowth = self$parGrowth,
  parMort = self$parMort,
  parReg = self$parReg,
  parGrowthEnv = self$parGrowthEnv,
  parMortEnv = self$parMortEnv,
  parRegEnv = self$parRegEnv,
  hidden_growth = self$hidden_growth,
  hidden_mort = self$hidden_mort,
  hidden_reg = self$hidden_reg,
  bias = self$bias,
  patch_size_ha = self$patch_size_ha,
  minLight = self$minLight,
  which = "all"
)
}
\arguments{
\item{sp}{integer Number of species.}

\item{device}{character Device to use ('cpu' or 'cuda').}

\item{parHeight}{torch.Tensor global parameters.}

\item{parGrowth}{torch.Tensor Growth parameters.}

\item{parMort}{torch.Tensor Mortality parameters.}

\item{parReg}{torch.Tensor Regeneration parameters.}

\item{parGrowthEnv}{torch.Tensor Growth environment parameters.}

\item{parMortEnv}{torch.Tensor Mortality environment parameters.}

\item{parRegEnv}{torch.Tensor Regeneration environment parameters.}

\item{hidden_growth}{list Hidden layers for growth model.}

\item{hidden_mort}{list Hidden layers for mortality model.}

\item{hidden_reg}{list Hidden layers for regeneration model.}
}
\description{
This function initializes the FINN model with specified species, device, global parameters, growth parameters, mortality parameters, regeneration parameters, and neural network settings.
}
