% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{initCohort}
\alias{initCohort}
\title{Initialize a CohortMat Instance}
\usage{
initCohort(
  obs_df = NULL,
  dbh = NULL,
  trees = NULL,
  species = NULL,
  dims = c(50, 30, 10),
  sp = 10,
  device = "cpu"
)
}
\arguments{
\item{obs_df}{A data frame containing columns "siteID", "patchID", "species", "dbh", and "trees". If provided, it will be used to initialize the tensors.}

\item{dbh}{A tensor or array representing the diameter at breast height. Defaults to \code{NULL}.}

\item{trees}{A tensor or array representing the number of trees. Defaults to \code{NULL}.}

\item{species}{A tensor or array representing the species. Defaults to \code{NULL}.}

\item{dims}{A numeric vector representing the dimensions of the arrays (sites, patches, cohorts). Defaults to \code{c(50, 30, 10)}.}

\item{sp}{An integer representing the number of species. Defaults to \code{10}.}

\item{device}{A character string specifying the device to use ('cpu' or 'cuda'). Defaults to \code{'cpu'}.}
}
\value{
A new instance of the \code{CohortMat} class.
}
\description{
A convenience function that calls \code{CohortMat$new} to initialize a new \code{CohortMat} object. This function has the same parameters and behavior as the \code{initialize} method of the \code{CohortMat} class.
}
\examples{
\dontrun{
# Initialize a CohortMat instance
cohort <- initCohort(obs_df = your_data_frame)
}
}
