% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_FINN}
\alias{install_FINN}
\title{Install FINN and its dependencies}
\usage{
install_FINN(
  conda = "auto",
  version = c("cpu", "gpu"),
  restart_session = TRUE,
  ...
)
}
\arguments{
\item{conda}{path to conda}

\item{version}{version = "cpu" for CPU version, or "gpu" for GPU version. (note MacOS users have to install 'cuda' binaries by themselves)}

\item{restart_session}{Restart R session after installing (note this will
only occur within RStudio).}

\item{...}{not supported}
}
\value{
No return value, called for side effects (installation of 'python' dependencies).
}
\description{
Install FINN and its dependencies
}
