% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_wrapper.R
\name{mort}
\alias{mort}
\title{Mortality function}
\usage{
mort(
  dbh,
  Species,
  nTree,
  parGlobal,
  parMort,
  pred,
  dtype = "float32",
  device = "cpu"
)
}
\arguments{
\item{dbh}{dbh of species, should be of dimensions [sites, patches, species]}

\item{Species}{species identifier, should be of dimensions [sites, patches, species]}

\item{nTree}{number of trees, should be of dimensions [sites, patches, species]}

\item{parGlobal}{species global parameter, should be of dimensions [species] (so it should be a vector)}

\item{parMort}{species mort parameter, should be of dimensions [species, 2]}

\item{pred}{predictions for environmental niche, should be of dimensions [sites, species]}

\item{dtype}{dtype}

\item{devie}{device}
}
\description{
Mortality function
}
\details{
if [length(dim) == 1] this dimension is assumed to be the species dimension (works only for R objects!)
}
\examples{
mort(rep(50, 5), 0:4, rep(5, 5), rep(1., 5), matrix(1., 5, 2), rep(0.5, 5))

}
