% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processes.R
\name{mortality_wo_growth}
\alias{mortality_wo_growth}
\title{mortality_wo_growth}
\usage{
mortality_wo_growth(
  dbh,
  species,
  trees,
  parMort,
  pred,
  light,
  base_steepness = 5,
  debug = F
)
}
\arguments{
\item{dbh}{dbh}

\item{species}{species}

\item{trees}{trees}

\item{parMort}{parMort}

\item{pred}{predictions}

\item{light}{available light}
}
\description{
mortality_wo_growth
}
