% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_setter.R, R/utils.R
\name{np_runif}
\alias{np_runif}
\title{Generate random numbers from a uniform distribution}
\usage{
np_runif(low, high, size)

np_runif(low, high, size)
}
\arguments{
\item{low}{numeric Lower bound of the uniform distribution.}

\item{high}{numeric Upper bound of the uniform distribution.}

\item{size}{numeric Size of the output array.}
}
\value{
array A numeric array of random numbers.

array A numeric array of random numbers.
}
\description{
This function generates random numbers from a uniform distribution with specified low and high values and size similar to np.random.uniform in Python.

This function generates random numbers from a uniform distribution with specified low and high values and size similar to np.random.uniform in Python.
}
\examples{
np_runif(0, 1, c(2, 3))

np_runif(0, 1, c(2, 3))

}
