% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pad_tensors_speed_up}
\alias{pad_tensors_speed_up}
\title{Padding tensors to remove dead cohorts}
\usage{
pad_tensors_speed_up(value, indices, org_dim, const = 0)
}
\description{
Padding tensors to remove dead cohorts
}
