% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINNInterface.R
\name{plot.finnModel}
\alias{plot.finnModel}
\title{finnModel plot}
\usage{
\method{plot}{finnModel}(x, plot = c("predictions", "loss"), ...)
}
\arguments{
\item{x}{a model fitted by \code{finnModel}}

\item{...}{ignored}
}
\value{
ggplot2 object
}
\description{
Plotting finnModel outputs (predictions or loss).
}
