% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINN.R
\name{predict.FINN}
\alias{predict.FINN}
\title{Predict from a fitted dnn model}
\usage{
\method{predict}{FINN}(
  object,
  newdata = NULL,
  type = c("link", "response", "class"),
  device = c("cpu", "cuda", "mps"),
  reduce = c("mean", "median", "none"),
  ...
)
}
\arguments{
\item{object}{a model created by \code{\link{FINN}}}

\item{newdata}{new data for predictions}

\item{device}{device on which network should be trained on.}

\item{...}{additional arguments passed to model$predict function}
}
\value{
prediction array
}
\description{
Predict from a fitted dnn model
}
