% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINNinterface.R
\name{predict.finnModel}
\alias{predict.finnModel}
\title{Predict Forest Dynamics}
\usage{
\method{predict}{finnModel}(object, init = NULL, env = NULL, ...)
}
\arguments{
\item{object}{object of class 'finnModel' created by \code{finn()}}

\item{init}{new \code{cohortMat$new()} object, if \code{NULL}, the initial cohortMat of the object is used}

\item{env}{new environment to simulate for}

\item{...}{currently ignored}
}
\description{
Predict Forest Dynamics
}
