% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINN.R
\name{predict}
\alias{predict}
\title{Predict the growth and mortality of trees BA_stemsed on the given inputs.}
\usage{
predict(
  dbh = NULL,
  trees = NULL,
  species = NULL,
  env = NULL,
  start_time = 1L,
  response = "dbh",
  pred_growth = NULL,
  pred_morth = NULL,
  pred_reg = NULL,
  patches = 50,
  debug = TRUE,
  y = NULL,
  c = NULL,
  update_step = 1L
)
}
\arguments{
\item{dbh}{torch.Tensor (Optional) The diameter at breast height of the trees. If NULL, it will be initialized using CohortMat.}

\item{trees}{torch.Tensor (Optional) The number of trees. If NULL, it will be initialized using CohortMat.}

\item{species}{torch.Tensor (Optional) The species of the trees. If NULL, it will be initialized using CohortMat.}

\item{env}{torch.Tensor The environmental data.}

\item{start_time}{integer The time at which to start recording the results.}

\item{response}{character The response variable to use for aggregating results. Can be "dbh", "BA_stem", or "BA_stem".}

\item{pred_growth}{torch.Tensor (Optional) The predicted growth values.}

\item{pred_morth}{torch.Tensor (Optional) The predicted mortality values.}

\item{pred_reg}{torch.Tensor (Optional) The predicted regeneration values.}

\item{patches}{float (Optional) The number of patches.}

\item{debug}{logical Whether to run in debug mode.}

\item{y}{response tensor}

\item{c}{number of tree tensor}

\item{update_step}{backpropagation step length}
}
\value{
list A list of predicted values for dbh and number of trees for the recorded time points.
}
\description{
This function predicts the growth and mortality of trees BA_stemsed on the given inputs, including diameter at breast height (dbh), number of trees, species, environmental data, and other parameters.
}
