% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINNInterface.R
\name{print.finnModel}
\alias{print.finnModel}
\title{Print finnModel}
\usage{
\method{print}{finnModel}(object, ...)
}
\arguments{
\item{object}{object of class finnModel}

\item{...}{not implemented}
}
\description{
print finnModel object created by \code{finn()}
}
