% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_wrapper.R
\name{reg}
\alias{reg}
\title{Reg function}
\usage{
reg(
  dbh,
  nTree,
  Species,
  parGlobal,
  parReg,
  pred,
  dtype = "float32",
  device = "cpu"
)
}
\arguments{
\item{dbh}{dbh of species, should be of dimensions [sites, patches, species, 1]}

\item{Species}{species identifier, should be of dimensions [sites, patches, species]}

\item{parGlobal}{species global parameter, should be of dimensions [species] (so it should be a vector)}

\item{parReg}{species growth parameter, should be of dimensions [species]}

\item{pred}{predictions for environmental niche, should be of dimensions [sites, species]}

\item{dtype}{dtype}

\item{devie}{device}
}
\description{
Reg function
}
\details{
if [length(dim) == 1] this dimension is assumed to be the species dimension (works only for R objects!)
}
\examples{
reg(rep(50, 5), 0:4, rep(1., 5), rep(1., 5), rep(0.5, 5))
}
