% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processes.R
\name{regeneration}
\alias{regeneration}
\title{Calculate the regeneration of forest patches based on the input parameters}
\usage{
regeneration(species, parReg, pred, light, debug = FALSE)
}
\arguments{
\item{species}{torch.Tensor species information.}

\item{parReg}{torch.Tensor Regeneration parameters. 0 <= parReg <= 1
This parameter denotes the fraction of light needed for a species to regenerate.
In general low values for high regeneration and high values for low regeneration.}

\item{pred}{torch.Tensor Prediction values.}

\item{light}{torch.Tensor Available light variable for calculation.}

\item{debug}{logical If TRUE, return the intermediate components as a list. Defaults to FALSE.}
}
\value{
torch.Tensor Regeneration values for forest patches.
}
\description{
This function calculates the regeneration of forest patches based on species information, regeneration parameters, prediction values, and available light.
}
