% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_poisson_relaxed}
\alias{sample_poisson_relaxed}
\title{Sample poisson relaxed}
\usage{
sample_poisson_relaxed(lmbd, num_samples = 50, temperature = 0.01)
}
\arguments{
\item{lmbd}{lambda}

\item{num_samples}{number of samples}

\item{temperature}{temperature}
}
\value{
A \code{torch} tensor of relaxed (differentiable) Poisson samples.
}
\description{
Sample poisson relaxed
}
\keyword{internal}
