% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finn-class.R
\name{simulateForest}
\alias{simulateForest}
\title{Simulate}
\usage{
simulateForest(
  model,
  env,
  disturbance = NULL,
  patches = 100L,
  patch_size = 0.1,
  init_cohort = NULL,
  batchsize = NULL,
  device = c("cpu", "gpu"),
  debug = FALSE
)
}
\arguments{
\item{model}{(\verb{ })\cr}

\item{env}{(\code{data.table|data.frame})\cr Data with environmental covariates must be passed as \code{data.table} or \code{data.frame}.}

\item{disturbance}{(\code{data.table|data.frame})\cr Data with disturbance rates must be passed as \code{data.table} or \code{data.frame}.}

\item{patches}{(\code{integer(1)})\cr Number of patches.}

\item{patch_size}{(\code{numeric(1)})\cr Patch size.}

\item{init_cohort}{(\code{CohortMat})\cr Initial cohort matrix of class \code{CohortMat}, created by \link{CohortMat}.}

\item{batchsize}{(\code{integer(1)})\cr Batch size, model will be trained in random batch sizes of the data to preserve memory and improve convergence.}

\item{device}{(\code{charachter(1)})\cr Should the model be fitted on the CPU or the GPU (Graphic card). Support is only for NVIDIA GPUs available.}

\item{debug}{(\code{logical(1)})\cr Debug modus or not. If \code{TRUE}, individual tree states are stored.}
}
\description{
Simulate
}
\details{
Simulate from a fitted FINN model
}
