% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINNInterface.R
\name{summary.finnModel}
\alias{summary.finnModel}
\title{Summary of finnModel}
\usage{
\method{summary}{finnModel}(object, ...)
}
\arguments{
\item{object}{object of class finnModel}

\item{...}{not implemented}
}
\description{
summary of finnModel object created by \code{finn()}
}
