% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{aggregate_results_old}
\alias{aggregate_results_old}
\title{Aggregate function}
\usage{
aggregate_results_old(
  labels,
  samples,
  Results,
  drop_rows = TRUE,
  sp_max = NULL
)
}
\arguments{
\item{labels}{labels}

\item{samples}{samples}

\item{Results}{Results}
}
\description{
Aggregate function
}
