% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xAI.R
\name{ALE}
\alias{ALE}
\title{Accumulated local effect plots}
\usage{
ALE(model, env, ...)
}
\arguments{
\item{model}{(\code{finn_class})\cr model object of class \code{finn_class}}

\item{env}{(\code{data.table|data.frame})\cr environmental covariates for which the ALE should be calculated}
}
\description{
Calculates accumulated local effects (ALE) for the three processes
}
