% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINN.R
\name{BA_stand}
\alias{BA_stand}
\title{Calculate the Basal Area of a Stand}
\usage{
BA_stand(dbh, trees, patch_size_ha)
}
\arguments{
\item{dbh}{A torch tensor or numeric vector representing the diameter at breast height of the trees in centimeters.}

\item{trees}{A torch tensor or numeric vector representing the number of trees.}

\item{patch_size_ha}{A numeric value representing the size of the patch in hectares.}
}
\value{
A numeric value representing the basal area of the stand in square meters per hectare.
}
\description{
This function calculates the basal area of a stand based on the diameter at breast height (dbh), the number of trees, and the patch size in hectares.
}
\details{
The basal area of a stand is the cross-sectional area of all trees in a stand per unit area. This function calculates the basal area per ha using the formula:
\deqn{BA = \left( \frac{\pi \left( \frac{\text{dbh}}{100} \right)^2}{4} \right) \times \text{trees} \div \text{patch\_size\_ha}}

The formula takes into account the diameter at breast height (dbh) in centimeters, the number of trees, and the size of the patch in hectares to calculate the basal area in square meters per hectare.

This plot illustrates the basal area for different combinations of dbh and number of trees.\if{html}{\out{
<img src="figures/BA_stand_plot2.png" alt="dbh, trees, basal area" style="max-width:70%;"/>
}}


Sensitivity of basal area for different combinations of dbh, number of trees to patch size.\if{html}{\out{
<img src="figures/BA_stand_plot1.png" alt="Patch size, dbh, trees, basal area" style="max-width:70%;"/>
}}
}
\examples{
# Example usage
dbh_vec <- seq(1, 200, 1)
trees_vec <- c(0:500, 10^(seq(2, 4, length.out = 20)))

# Generate test data for a patch size of 0.1
patch_size <- 0.1
cohort_df1 <- expand.grid(
  trees_ha = trees_vec,
  patch_size_ha = patch_size,
  dbh = dbh_vec
)

cohort_df1 <- data.frame(
  patchID = 1,
  cohortID = 1,
  species = 1,
  cohort_df1
)

cohort_df1$siteID <- 1:nrow(cohort_df1)
cohort_df1$trees <- round(cohort_df1$trees_ha * patch_size)

cohort <- CohortMat$new(obs_df = cohort_df1)

cohort_df1$basal_area <- torch::as_array(BA_stand(cohort$dbh, cohort$trees, patch_size_ha = patch_size))

# View the first few rows of the resulting data frame
head(cohort_df1)

# Basic plot showing the function of trees and dbh for basal area

# only keep rows with basal area <100
cohort_df1 <- cohort_df1[cohort_df1$basal_area < 100,]

library(ggplot2)
ggplot(cohort_df1, aes(x = dbh, y = basal_area, color = trees, group = trees)) +
  geom_line() +
  ylab("Basal Area (m^2/ha)") +
  xlab("Diameter at Breast Height (cm)") +
  scale_color_viridis_c(name = "Trees per ha", trans = "log10", option = "magma", direction = -1) +
  ggtitle("Basal Area as a Function of Trees and DBH")
}
