% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processes.R
\name{BA_stem}
\alias{BA_stem}
\title{Calculate the basal area of a tree given the diameter at breast height (dbh).}
\usage{
BA_stem(dbh)
}
\arguments{
\item{dbh}{torch.Tensor The diameter at breast height of the tree.}
}
\value{
torch.Tensor The basal area of the tree.
}
\description{
This function calculates the basal area of a tree given the diameter at breast height (dbh).
}
\examples{
\dontrun{
dbh = torch::torch_tensor(50)
basal_area = BA_stem(dbh)
print(basal_area)
}

}
