% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processes.R
\name{BA_stem}
\alias{BA_stem}
\title{Calculate the BA_stemsal area of a tree given the diameter at breast height (dbh).}
\usage{
BA_stem(dbh)
}
\arguments{
\item{dbh}{torch.Tensor The diameter at breast height of the tree.}
}
\value{
torch.Tensor The BA_stemsal area of the tree.
}
\description{
This function calculates the BA_stemsal area of a tree given the diameter at breast height (dbh).
}
\examples{
dbh = torch::torch_tensor(50)
BA_stemsal_area = BA_stem(dbh)
print(BA_stemsal_area)

}
