% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{binomial_from_gamma}
\alias{binomial_from_gamma}
\title{Sample from binomial with gradient}
\usage{
binomial_from_gamma(n, p, sample_size = 1)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of success}

\item{sample_size}{sample size}
}
\description{
Sample from binomial with gradient
}
