% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climateDF2array.R
\name{climateDF2array}
\alias{climateDF2array}
\title{Transform Climate Data Table to Arrays}
\usage{
climateDF2array(climate_dt, include_month = FALSE)
}
\arguments{
\item{climate_dt}{data.frame The climate data table containing year, month, uniquePLOTid, and environmental variables (e.g., tmp, pre) columns.}

\item{include_month}{logical Whether to include the month dimension in the output array. Defaults to FALSE.}
}
\value{
An array with dimensions for sites, year, and Nenv (or optionally, sites, year, month, and Nenv).
}
\description{
This function transforms a climate data table into an array with dimensions for sites, year, and Nenv.
Optionally, it can create an array with dimensions for sites, year, month, and Nenv.
}
\examples{
climate_dt <- data.frame(
  year = rep(1960:1961, each = 2),
  month = rep(1:2, times = 2),
  uniquePLOTid = 1:4,
  tmp = runif(4, -10, 30),
  pre = runif(4, 0, 200)
)
result <- climateDF2array(climate_dt, include_month = TRUE)

}
