% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortMat.R
\name{CohortMat}
\alias{CohortMat}
\title{Cohort Matrix Class}
\description{
An R6 class for managing cohorts of trees in forest models. This class allows for the initialization, transformation, and manipulation of cohorts represented by arrays of dbh, trees, and species.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dbh}}{A tensor representing the diameter at breast height for each cohort.}

\item{\code{trees}}{A tensor representing the number of trees in each cohort.}

\item{\code{species}}{A tensor representing the species of each cohort.}

\item{\code{dims}}{A vector representing the dimensions of the arrays (sites, patches, cohorts).}

\item{\code{sp}}{An integer representing the number of species.}

\item{\code{device}}{A character string specifying the device to use ('cpu' or 'cuda').}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CohortMat-obsDF2arrays}{\code{CohortMat$obsDF2arrays()}}
\item \href{#method-CohortMat-new}{\code{CohortMat$new()}}
\item \href{#method-CohortMat-asDF}{\code{CohortMat$asDF()}}
\item \href{#method-CohortMat-clone}{\code{CohortMat$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortMat-obsDF2arrays"></a>}}
\if{latex}{\out{\hypertarget{method-CohortMat-obsDF2arrays}{}}}
\subsection{Method \code{obsDF2arrays()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortMat$obsDF2arrays(obs_dt)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortMat-new"></a>}}
\if{latex}{\out{\hypertarget{method-CohortMat-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortMat$new(
  dbh = self$dbh,
  trees = self$trees,
  species = self$species,
  dims = self$dims,
  sp = self$sp,
  device = self$device,
  obs_df = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortMat-asDF"></a>}}
\if{latex}{\out{\hypertarget{method-CohortMat-asDF}{}}}
\subsection{Method \code{asDF()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortMat$asDF()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortMat-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CohortMat-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortMat$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
