% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_wrapper.R
\name{competition}
\alias{competition}
\title{Competition function}
\usage{
competition(
  dbh,
  nTree,
  Species,
  parGlobal,
  h = NULL,
  minLight = 50,
  dtype = "float32",
  device = "cpu"
)
}
\arguments{
\item{dbh}{dbh of species, should be of dimensions [sites, patches, species, 1]}

\item{nTree}{number of trees, should be of dimensions [sites, patches, species]}

\item{Species}{species identifier, should be of dimensions [sites, patches, species]}

\item{parGlobal}{species global parameter, should be of dimensions [species] (so it should be a vector)}

\item{h}{height of species, if not null, should be [1,1]}

\item{minLight}{minimal light}

\item{dtype}{dtype}

\item{devie}{device}
}
\description{
Competition function
}
\details{
if [length(dim) == 1] this dimension is assumed to be the species dimension (works only for R objects!)
}
\examples{
competition(rep(50, 5), 0:4, rep(1., 5))
}
