% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINN.R
\name{competition}
\alias{competition}
\title{Compute the fraction of available light (light) for each cohort BA_stemsed on the given parameters.}
\usage{
competition(
  dbh,
  species,
  trees,
  parHeight,
  h = NULL,
  minLight = 50,
  path_size_ha = 0.1
)
}
\arguments{
\item{dbh}{torch.Tensor Diameter at breast height for each cohort.}

\item{species}{torch.Tensor species index for each cohort.}

\item{trees}{Number of trees.}

\item{parHeight}{torch.Tensor global parameters for all species.}

\item{h}{torch.Tensor (Optional) Height of each cohort. Defaults to NULL.}

\item{minLight}{float (Optional) Minimum light requirement. Defaults to 50.}
}
\value{
torch.Tensor Fraction of available light (light) for each cohort.
}
\description{
This function calculates the fraction of available light for each cohort of trees BA_stemsed on their diameter at breast height (dbh), species, number of trees, and global parameters.
}
\examples{
competition(dbh = torch::torch_tensor(c(10, 15, 20)), species = torch::torch_tensor(c(1, 2, 1)),
        trees = 100, parHeight = torch::torch_tensor(c(0.3, 0.5)), h = torch::torch_tensor(c(5, 7, 6)), minLight = 40)
}
