% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processes_utils.R
\name{createHybrid}
\alias{createHybrid}
\title{Create a hybrid model Object}
\usage{
createHybrid(
  formula = NULL,
  hidden = NULL,
  optimize = TRUE,
  dispersion_parameter = 1,
  NN = NULL,
  dropout = 0,
  encoder_layers = 1L
)
}
\arguments{
\item{formula}{An optional formula specifying the structure of the model. Default is \code{NULL}, which results in the formula \code{~.} being used.}

\item{hidden}{A list specifying the hidden layers for neural network models. Default is an empty list.}

\item{optimize}{Logical indicating whether the process function should be optimized. Default is \code{FALSE}.}

\item{dispersion_parameter}{init dispersion parameter, if available (currently only supported for regeneration rate that is based on a negative binomial).}

\item{NN}{pass custom NN to model}

\item{dropout}{dropout rate of neural networks}
}
\value{
A list of class "process" containing the process definition and associated parameters.
}
\description{
This function creates a process object that is used to define a specific ecological process (e.g., growth, mortality, regeneration) within a forest simulation model. The process object can include custom functions, species parameters, environmental initialization, and hidden layers for neural networks.
}
\examples{
growth_process <- createProcess(formula = ~temperature + precipitation, func = growthFunction)

}
