% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finn-class.R
\name{finn}
\alias{finn}
\title{Forest Informed Neural Network}
\usage{
finn(
  N_species,
  mortality_process = NULL,
  growth_process = NULL,
  regeneration_process = NULL,
  competition_process = NULL,
  recruits_dbh = 1
)
}
\arguments{
\item{N_species}{(\code{integer(1)})\cr Number of species.}

\item{mortality_process}{(\code{function})\cr Mortality function.}

\item{growth_process}{(\code{function})\cr Growth function.}

\item{regeneration_process}{(\code{function})\cr Regeneration function.}

\item{competition_process}{(\code{function})\cr Competition function.}

\item{recruits_dbh}{(\code{numeric(1)})\cr Starting dbh for recruits. Has value 1.0 as default.}
}
\description{
Creates a Forest Informed Neural Network (FINN) object. Environmental components or the entire process can be replaced by deep neural networks.
}
