% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINN.R
\name{FINN}
\alias{FINN}
\title{FINN: Forest Inventory Neural Network}
\description{
A class to initialize and train a Forest Inventory Neural Network (FINN) for predicting tree growth, mortality, and regeneration.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{np_runif}}{Function to generate random numbers from a uniform distribution similar to np.random.uniform in Python.}

\item{\code{sp}}{integer Number of species.}

\item{\code{device}}{character Device to use ('cpu' or 'cuda').}

\item{\code{parHeight}}{torch.Tensor global parameters.}

\item{\code{parGrowth}}{torch.Tensor Growth parameters.}

\item{\code{parMort}}{torch.Tensor Mortality parameters.}

\item{\code{parReg}}{torch.Tensor Regeneration parameters.}

\item{\code{parGrowthEnv}}{torch.Tensor Growth environment parameters.}

\item{\code{parMortEnv}}{torch.Tensor Mortality environment parameters.}

\item{\code{parRegEnv}}{torch.Tensor Regeneration environment parameters.}

\item{\code{hidden_growth}}{list Hidden layers for growth model.}

\item{\code{hidden_mort}}{list Hidden layers for mortality model.}

\item{\code{hidden_reg}}{list Hidden layers for regeneration model.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{np_runif}}{Function to generate random numbers from a uniform distribution similar to np.random.uniform in Python.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FINN-new}{\code{FINN$new()}}
\item \href{#method-FINN-build_NN}{\code{FINN$build_NN()}}
\item \href{#method-FINN-predict}{\code{FINN$predict()}}
\item \href{#method-FINN-fit}{\code{FINN$fit()}}
\item \href{#method-FINN-np_runif}{\code{FINN$np_runif()}}
\item \href{#method-FINN-set_weights_nnGrowthEnv}{\code{FINN$set_weights_nnGrowthEnv()}}
\item \href{#method-FINN-set_weights_nnMortEnv}{\code{FINN$set_weights_nnMortEnv()}}
\item \href{#method-FINN-set_weights_nnRegEnv}{\code{FINN$set_weights_nnRegEnv()}}
\item \href{#method-FINN-get_parMort}{\code{FINN$get_parMort()}}
\item \href{#method-FINN-get_parGrowth}{\code{FINN$get_parGrowth()}}
\item \href{#method-FINN-get_parHeight}{\code{FINN$get_parHeight()}}
\item \href{#method-FINN-get_parReg}{\code{FINN$get_parReg()}}
\item \href{#method-FINN-set_parMort}{\code{FINN$set_parMort()}}
\item \href{#method-FINN-set_parGrowth}{\code{FINN$set_parGrowth()}}
\item \href{#method-FINN-set_parHeight}{\code{FINN$set_parHeight()}}
\item \href{#method-FINN-set_parReg}{\code{FINN$set_parReg()}}
\item \href{#method-FINN-clone}{\code{FINN$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FINN-new"></a>}}
\if{latex}{\out{\hypertarget{method-FINN-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FINN$new(
  sp = self$sp,
  env = self$env,
  device = self$device,
  parHeight = self$parHeight,
  parGrowth = self$parGrowth,
  parMort = self$parMort,
  parReg = self$parReg,
  parGrowthEnv = self$parGrowthEnv,
  parMortEnv = self$parMortEnv,
  parRegEnv = self$parRegEnv,
  hidden_growth = self$hidden_growth,
  hidden_mort = self$hidden_mort,
  hidden_reg = self$hidden_reg,
  bias = self$bias,
  patch_size_ha = self$patch_size_ha,
  minLight = self$minLight,
  disturbance = self$disturbance,
  which = c("all", "env", "species")
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FINN-build_NN"></a>}}
\if{latex}{\out{\hypertarget{method-FINN-build_NN}{}}}
\subsection{Method \code{build_NN()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FINN$build_NN(
  self,
  input_shape,
  output_shape,
  hidden,
  bias,
  activation,
  dropout,
  last_activation = "sigmoid"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FINN-predict"></a>}}
\if{latex}{\out{\hypertarget{method-FINN-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FINN$predict(
  dbh = NULL,
  trees = NULL,
  species = NULL,
  env = NULL,
  start_time = 1L,
  response = "dbh",
  pred_growth = NULL,
  pred_morth = NULL,
  pred_reg = NULL,
  patches = 50,
  debug = TRUE,
  y = NULL,
  c = NULL,
  update_step = 1L
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FINN-fit"></a>}}
\if{latex}{\out{\hypertarget{method-FINN-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FINN$fit(
  X = NULL,
  Y = NULL,
  initCohort = NULL,
  epochs = 2L,
  batch_size = 20L,
  learning_rate = 0.1,
  start_time = 1,
  patches = 50L,
  response = "dbh",
  update_step = 1L
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FINN-np_runif"></a>}}
\if{latex}{\out{\hypertarget{method-FINN-np_runif}{}}}
\subsection{Method \code{np_runif()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FINN$np_runif(low, high, size)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FINN-set_weights_nnGrowthEnv"></a>}}
\if{latex}{\out{\hypertarget{method-FINN-set_weights_nnGrowthEnv}{}}}
\subsection{Method \code{set_weights_nnGrowthEnv()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FINN$set_weights_nnGrowthEnv(weights)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FINN-set_weights_nnMortEnv"></a>}}
\if{latex}{\out{\hypertarget{method-FINN-set_weights_nnMortEnv}{}}}
\subsection{Method \code{set_weights_nnMortEnv()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FINN$set_weights_nnMortEnv(weights)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FINN-set_weights_nnRegEnv"></a>}}
\if{latex}{\out{\hypertarget{method-FINN-set_weights_nnRegEnv}{}}}
\subsection{Method \code{set_weights_nnRegEnv()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FINN$set_weights_nnRegEnv(weights)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FINN-get_parMort"></a>}}
\if{latex}{\out{\hypertarget{method-FINN-get_parMort}{}}}
\subsection{Method \code{get_parMort()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FINN$get_parMort()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FINN-get_parGrowth"></a>}}
\if{latex}{\out{\hypertarget{method-FINN-get_parGrowth}{}}}
\subsection{Method \code{get_parGrowth()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FINN$get_parGrowth()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FINN-get_parHeight"></a>}}
\if{latex}{\out{\hypertarget{method-FINN-get_parHeight}{}}}
\subsection{Method \code{get_parHeight()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FINN$get_parHeight()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FINN-get_parReg"></a>}}
\if{latex}{\out{\hypertarget{method-FINN-get_parReg}{}}}
\subsection{Method \code{get_parReg()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FINN$get_parReg()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FINN-set_parMort"></a>}}
\if{latex}{\out{\hypertarget{method-FINN-set_parMort}{}}}
\subsection{Method \code{set_parMort()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FINN$set_parMort(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FINN-set_parGrowth"></a>}}
\if{latex}{\out{\hypertarget{method-FINN-set_parGrowth}{}}}
\subsection{Method \code{set_parGrowth()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FINN$set_parGrowth(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FINN-set_parHeight"></a>}}
\if{latex}{\out{\hypertarget{method-FINN-set_parHeight}{}}}
\subsection{Method \code{set_parHeight()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FINN$set_parHeight(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FINN-set_parReg"></a>}}
\if{latex}{\out{\hypertarget{method-FINN-set_parReg}{}}}
\subsection{Method \code{set_parReg()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FINN$set_parReg(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FINN-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FINN-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FINN$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
