% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finn-class.R
\name{finn}
\alias{finn}
\title{finn module}
\usage{
finn(
  N_species,
  mortality_process = NULL,
  growth_process = NULL,
  regeneration_process = NULL,
  competition_process = NULL,
  recruits_dbh = 1
)
}
\arguments{
\item{dbh}{torch.Tensor (Optional). Diameter at breast height of the trees.}

\item{trees}{torch.Tensor (Optional). Number of trees.}

\item{species}{torch.Tensor (Optional). Species of the trees.}

\item{env}{torch.Tensor. Environmental data.}

\item{disturbance}{torch.Tensor. Disturbance rates.}

\item{start_time}{integer. Time at which to start recording the results.}

\item{pred_growth}{torch.Tensor (Optional). Predicted growth values.}

\item{pred_mort}{torch.Tensor (Optional). Predicted mortality values.}

\item{pred_reg}{torch.Tensor (Optional). Predicted regeneration values.}

\item{patches}{numeric. Number of patches.}

\item{debug}{logical. Run in debug mode if TRUE.}

\item{y}{torch.Tensor. Response tensor for target data.}

\item{c}{torch.Tensor. Number of tree tensor.}

\item{update_step}{integer. Backpropagation step length.}

\item{verbose}{logical. Print progress if TRUE.}

\item{weights}{weights for reweighting the loss}

\item{year_sequence}{at which year indices should the predictions compared with the observed values}
}
\value{
list. A list of predicted values for dbh, number of trees, and other recorded time points. If \code{debug} is TRUE, raw results and cohorts are also returned.
}
\description{
Predicts the growth, mortality, and regeneration of trees based on the given inputs.

The \code{forward} method generates predictions for tree growth, mortality, and regeneration for the specified species across different environmental conditions. It uses the initialized model parameters and can handle optional input tensors like diameter at breast height (dbh), number of trees, and species. If these are not provided, they will be initialized internally.
}
