% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINN.R
\name{FINN}
\alias{FINN}
\title{FINN model}
\usage{
FINN(
  formula,
  data = NULL,
  response = NULL,
  patches = 10L,
  epochs = 1L,
  learning_rate = 1,
  batch_size = 70L,
  device = "cpu",
  start_time = 0.7,
  X = NULL,
  Y = NULL,
  init_global = NULL,
  init_growth = NULL,
  init_mort = NULL,
  init_reg = NULL,
  init_growth_env = NULL,
  init_mort_env = NULL,
  init_reg_env = NULL,
  hidden_growth = list(),
  hidden_mort = list(),
  hidden_reg = list(),
  simulate = FALSE
)
}
\arguments{
\item{formula}{formula}

\item{data}{environmental data, must contain the variables timestep and site}

\item{response}{species dbh and number of trees, must contain the variables timestep, site, nTree, and Species}

\item{patches}{number of patches}

\item{epochs}{number of epochs (optimization)}

\item{learning_rate}{learning rate for stochastic gradient descent}

\item{batch_size}{batch_size for stochastic gradient descent}

\item{device}{cpu or cuda, if available}

\item{start_time}{when to record, fractional, from 0 to 1}

\item{X}{alternative parametrization, dims: (site, time, predictors)}

\item{Y}{alternative parametrization, dims: (site, time, species, 2), 2 for dbh/ba/ba*t and number of Trees (must be integers)}

\item{init_global}{Initial values for global parameters (optional).}

\item{init_growth}{Initial values for growth parameters (optional).}

\item{init_mort}{Initial values for mortality parameters (optional).}

\item{init_reg}{Initial values for regeneration parameters (optional).}

\item{init_growth_env}{Initial values for NN env, must be a list of matrices (each matrix == one hidden layer)}

\item{init_mort_env}{Initial values for NN env, must be a list of matrices (each matrix == one hidden layer)}

\item{init_reg_env}{Initial values for NN env, must be a list of matrices (each matrix == one hidden layer)}

\item{hidden_growth}{List of specifications for hidden layers in growth NN (optional).}

\item{hidden_mort}{List of specifications for hidden layers in mort NN (optional).}

\item{hidden_reg}{List of specifications for hidden layers in reg NN (optional)}

\item{simulate}{fit or simulation modus, if simulation, model is not fitted and will be returned}
}
\value{
An object of class `FINN` containing the fitted model, data, and functions to access the model and predictions.
}
\description{
FINN model
}
\author{
Maximilian Pichler
}
