% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINN.R
\name{FINN}
\alias{FINN}
\title{FINN model}
\usage{
FINN(
  formula,
  data = NULL,
  response = NULL,
  patches = 10L,
  epochs = 1L,
  learning_rate = 1,
  batch_size = 70L,
  device = "cpu",
  start_time = 0.7,
  X = NULL,
  Y = NULL
)
}
\arguments{
\item{formula}{formula}

\item{data}{environmental data, must contain the variables timestep and site}

\item{response}{species dbh and number of trees, must contain the variables timestep, site, nTree, and Species}

\item{patches}{number of patches}

\item{epochs}{number of epochs (optimization)}

\item{learning_rate}{learning rate for stochastic gradient descent}

\item{batch_size}{batch_size for stochastic gradient descent}

\item{device}{cpu or cuda, if available}

\item{start_time}{when to record, fractional, from 0 to 1}

\item{X}{alternative parametrization, dims: (site, time, predictors)}

\item{Y}{alternative parametrization, dims: (site, time, species, 2), 2 for dbh/ba/ba*t and number of Trees (must be integers)}
}
\description{
FINN model
}
\author{
Maximilian Pichler
}
