% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINNinterface.R
\name{finn}
\alias{finn}
\title{Forest informed neural network}
\usage{
finn(
  data = NULL,
  env,
  mortalityProcess = NULL,
  growthProcess = NULL,
  regenerationProcess = NULL,
  competitionProcess = NULL,
  height = NULL,
  patches = 10L,
  patch_size = 0.1,
  init = NULL,
  batchsize = 50L,
  epochs = 20L,
  lr = 0.1,
  device = c("cpu", "gpu"),
  bootstrap = NULL,
  parallel = FALSE,
  NGPU = 1,
  ...
)
}
\arguments{
\item{data}{data containing species information (dbh, ba, trees, AL, growth, mortality rates, and regeneration rates)}

\item{env}{A data frame containing the environmental variables used in the simulation.}

\item{height}{A function or set of parameters defining the height-growth relationship, if applicable.}

\item{patches}{An integer specifying the number of patches in the simulation. Default is 10.}

\item{patch_size}{A numeric value representing the size of each patch. Default is 0.1.}

\item{init}{A custom initialization object for the simulation, if available.}

\item{device}{A character string specifying whether to use \code{"cpu"} or \code{"gpu"} for computation. Default is \code{"cpu"}.}

\item{bootstrap}{bootstrap model or not (computationally expensive!)}

\item{parallel}{A logical value indicating whether to run the bootstrapping in parallel. Default is \code{FALSE}.}

\item{NGPU}{An integer specifying the number of GPUs to use if \code{device = "gpu"}. Default is 1 (only used for parallel bootstrapping).}

\item{...}{arguments passed to \code{simulateForest()}}

\item{mortality}{A process object or formula defining the mortality process.}

\item{growth}{A process object or formula defining the growth process.}

\item{regeneration}{A process object or formula defining the regeneration process.}

\item{competition}{A function defining the competition process, if applicable.}

\item{sp}{An integer specifying the number of species in the simulation. Default is 5.}
}
\value{
A list of predicted values for each patch and species, containing arrays of tree diameters and tree counts.
}
\description{
This function simulates forest dynamics, including mortality, growth, regeneration, and competition,
across multiple patches and species.
}
\examples{
env_data <- data.frame(siteID = 1:5, year = 2001:2005, temp = rnorm(5), precip = rnorm(5))
predictions <- simulateForest(env = env_data, patches = 3, sp = 2)

}
