% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finn-class.R
\name{fit}
\alias{fit}
\title{Fit FINN}
\usage{
fit(
  model,
  data = NULL,
  env,
  disturbance = NULL,
  patches = 100L,
  patch_size = 0.1,
  init_cohort = NULL,
  epochs = 20L,
  lr = 0.01,
  loss = c(dbh = "mse", ba = "mse", trees = "poisson", growth = "mse", mortality = "mse",
    regeneration = "nbinom"),
  weights = rep(1, 6),
  optimizer = torch::optim_ignite_adam,
  batchsize = NULL,
  device = c("cpu", "gpu"),
  update_step = 1L,
  start_time = 1L,
  plot_progress = TRUE,
  folder = NULL,
  checkpoints = 100L,
  shuffle = TRUE,
  record_gradients = FALSE
)
}
\arguments{
\item{model}{(\code{finn_class})\cr Object of class \code{finn_class} created by \link{finn}.}

\item{data}{(\code{data.table|data.frame})\cr Data about demographic rates and stand variables must be passed as \code{data.table} or \code{data.frame}.}

\item{env}{(\code{data.table|data.frame})\cr Data with environmental covariates must be passed as \code{data.table} or \code{data.frame}.}

\item{disturbance}{(\code{data.table|data.frame})\cr Data with disturbance rates must be passed as \code{data.table} or \code{data.frame}.}

\item{patches}{(\code{integer(1)})\cr Number of patches.}

\item{patch_size}{(\code{numeric(1)})\cr Patch size.}

\item{init_cohort}{(\code{CohortMat})\cr Initial cohort matrix of class \code{CohortMat}, created by \link{CohortMat}}

\item{epochs}{(\code{integer(1)})\cr Number of iteration steps.}

\item{lr}{(\code{numeric(1)})\cr Learning rate of the optimizer.}

\item{loss}{(\code{character(6)})\cr Named vector of the different losses. Names should be \code{dbh}, \code{ba}, \code{trees}, \code{growth}, \code{mortality}, and \code{regeneration}. Supported losses are \code{mse}, \code{poisson}, \code{nbinom}, and \code{gaussian}.}

\item{weights}{(\code{numeric(6)})\cr Weights of the different losses.}

\item{optimizer}{(\code{torch_optimizer_generator})\cr Optimizer from the \code{torch} package.}

\item{batchsize}{(\code{integer(1)})\cr Batch size, model will be trained in random batch sizes of the data to preserve memory and improve convergence.}

\item{device}{(\code{charachter(1)})\cr Should the model be fitted on the CPU or the GPU (Graphic card). Support is only for NVIDIA GPUs available.}

\item{update_step}{(\code{integer(1)})\cr Number of steps for which the gradient should be calculated. Automatic differentation becomes slow for larger update steps and the risk of vanishing gradients increases.}

\item{start_time}{(\code{integer(1)})\cr Starting from which year should the model be fitted. Can be used to use on burn-in.}

\item{plot_progress}{(\code{logical(1)})\cr Plot fitting progress (losses) or not.}

\item{folder}{(\code{character(1)})\cr Path to folder for saving checkpoint models. If \code{NULL}, no models will be saved during the training.}

\item{checkpoints}{(\code{integer(1)})\cr Interval size in epochs for saving checkpoint models.}

\item{shuffle}{(\code{logical(1)})\cr Shuffle data or not.}

\item{record_gradients}{(\code{logical(1)})\cr Record the gradients of all parameters or not. Can get large for many epochs.}
}
\description{
Fit (calibrate) the FINN model.
}
