% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINN.R
\name{fit}
\alias{fit}
\title{Fit the model to the given data.}
\usage{
fit(
  X = NULL,
  Y = NULL,
  initCohort = NULL,
  epochs = 2L,
  batch_size = 20L,
  learning_rate = 0.1,
  start_time = 1,
  patches = 50L,
  response = "dbh",
  update_step = 1L
)
}
\arguments{
\item{X}{torch.Tensor (Optional) The input data. Default is NULL.}

\item{Y}{torch.Tensor (Optional) The target data. Default is NULL.}

\item{initCohort}{list (Optional) Initial cohort data. Default is NULL.}

\item{epochs}{integer The number of epochs to train the model. Default is 2.}

\item{batch_size}{integer The BA_stemtch size for training. Default is 20.}

\item{learning_rate}{float The learning rate for the optimizer. Default is 0.1.}

\item{start_time}{float The start time for prediction. Default is 0.5.}

\item{patches}{integer The number of patches. Default is 50.}

\item{response}{character The response variable. Default is "dbh", other options are 'BA_stem' or 'BA_stem*nT'.}

\item{update_step}{backpropagation step length}
}
\value{
None
}
\description{
This function fits the model to the given data using specified epochs, BA_stemtch size, learning rate, start time, and response variable.
}
