% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_wrapper.R
\name{growth}
\alias{growth}
\title{Growth function}
\usage{
growth(
  dbh,
  Species,
  parGlobal,
  parGrowth,
  pred,
  dtype = "float32",
  device = "cpu"
)
}
\arguments{
\item{dbh}{dbh of species, should be of dimensions [sites, patches, species, 1]}

\item{Species}{species identifier, should be of dimensions [sites, patches, species]}

\item{parGlobal}{species global parameter, should be of dimensions [species] (so it should be a vector)}

\item{parGrowth}{species growth parameter, should be of dimensions [species, 2]}

\item{pred}{predictions for environmental niche, should be of dimensions [sites, species]}

\item{dtype}{dtype}

\item{devie}{device}
}
\description{
Growth function
}
\details{
if [length(dim) == 1] this dimension is assumed to be the species dimension (works only for R objects!)
}
\examples{
growth(rep(50, 5), 0:4, rep(1., 5), matrix(1., 5, 2), rep(0.5, 5))
}
