% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processes.R
\name{growth}
\alias{growth}
\title{Calculate growth}
\usage{
growth(
  dbh,
  species,
  parGrowth,
  parMort,
  pred,
  light,
  light_steepness = 10,
  debug = F
)
}
\arguments{
\item{dbh}{torch.Tensor Diameter at breast height.}

\item{species}{torch.Tensor species of tree.}

\item{parGrowth}{torch.Tensor Growth parameters.}

\item{parMort}{torch.Tensor Mortality parameters.}

\item{pred}{torch.Tensor Predicted values.}

\item{light}{torch.Tensor Accumulated Light.}
}
\value{
torch.Tensor A tensor representing the forest plot growth.
}
\description{
This function calculates growth BA_stemsed on specified parameters.
}
