% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINN.R
\name{height}
\alias{height}
\title{Calculate the height of a tree based on its diameter at breast height and an alometry parameter.}
\usage{
height(dbh, parHeight)
}
\arguments{
\item{dbh}{A numeric value representing the diameter at breast height of the tree in cm.}

\item{parHeight}{A numeric value representing the species height alometry.}
}
\value{
A numeric value representing the calculated height of the tree.
}
\description{
Calculate the height of a tree based on its diameter at breast height and an alometry parameter.
}
\details{
This function calculates the height of a tree based on the diameter at breast height (dbh) and a parameter parHeight.

The height is calculated using the formula:
\deqn{height = \left( \exp \left( \frac{(\text{dbh} \times \text{parHeight})}{(\text{dbh} + 100)} \right) - 1 \right) \times 100 + 0.001}
where dbh is the diameter at breast height of the tree in cm and parHeight is an alometric species specific parameter.

All parameters of parHeight from 0 to 1 result in physiologicaly plausible heights.
The range from 0.3 to 0.9 results in realistic tree heights.
Values of parHeight close to 1 are physiologically almost impossible, below 0.3 is suitable for small tree species and shrubs.\if{html}{\out{
<br>
<img src="figures/height_plot1.png" alt="Parameter range" style="max-width:70%;"/>
}}
}
\examples{
height(30, 0.5)
height(c(30), c(0.5,0.3))
height(c(30,20), c(0.5))

}
