% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processes.R
\name{mortality}
\alias{mortality}
\title{Mortality}
\usage{
mortality(
  dbh,
  species,
  trees,
  parMort,
  pred,
  light,
  base_steepness = 5,
  debug = F,
  growth = NULL
)
}
\arguments{
\item{dbh}{dbh}

\item{species}{species}

\item{trees}{trees}

\item{parMort}{parMort}

\item{pred}{predictions}

\item{light}{available light}
}
\description{
Mortality
}
