% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINN.R
\name{mortality}
\alias{mortality}
\title{Mortality}
\usage{
mortality(dbh, species, trees, parMort, pred, light)
}
\arguments{
\item{dbh}{dbh}

\item{species}{species}

\item{trees}{trees}

\item{parMort}{parMort}

\item{pred}{predictions}

\item{light}{available light}
}
\description{
Mortality
}
