% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortMat.R
\name{obsDF2arrays}
\alias{obsDF2arrays}
\title{Convert observation data frame to arrays}
\usage{
obsDF2arrays(obs_dt)
}
\arguments{
\item{obs_dt}{data.frame The observation data table containing siteID, patchID, cohortID, species, dbh, and trees columns.}
}
\value{
A list of arrays for species, dbh, and trees
}
\description{
Convert observation data frame to arrays
}
\examples{
obs_dt <- data.frame(siteID = c(1, 1, 2), patchID = c(1, 2, 1), cohortID = c(1, 1, 2), species = c("A", "B", "A"), dbh = c(10, 20, 30), trees = c(100, 200, 150))
result <- obsDF2arrays(obs_dt)
}
