% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortMat.R
\name{obsDF2arrays}
\alias{obsDF2arrays}
\title{Transform Observation Data Table to Arrays}
\usage{
obsDF2arrays(obs_dt)
}
\arguments{
\item{obs_dt}{data.table The observation data table containing siteID, patchID, cohortID, species, dbh, and trees columns.}
}
\value{
A list containing three arrays: species, dbh, and trees.
}
\description{
This function transforms an observation data table into three arrays: species, dbh, and trees.
}
\examples{
obs_dt <- data.table(siteID = c(1, 1, 2), patchID = c(1, 2, 1), cohortID = c(1, 1, 2), species = c("A", "B", "A"), dbh = c(10, 20, 30), trees = c(100, 200, 150))
result <- obsDF2arrays(obs_dt)

}
