% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finn-class.R
\name{predict.finn_class}
\alias{predict.finn_class}
\title{Predict from a FINN model}
\usage{
\method{predict}{finn_class}(
  object,
  env,
  disturbance = NULL,
  patches = 100L,
  patch_size = 0.1,
  init_cohort = NULL,
  batchsize = NULL,
  device = c("cpu", "gpu"),
  debug = FALSE,
  ...
)
}
\arguments{
\item{object}{(\code{finn_class})\cr Object of class \code{finn_class} created by \link{finn}.}

\item{env}{(\code{data.table|data.frame})\cr Data with environmental covariates must be passed as \code{data.table} or \code{data.frame}.}

\item{disturbance}{(\code{data.table|data.frame})\cr Data with disturbance rates must be passed as \code{data.table} or \code{data.frame}.}

\item{patches}{(\code{integer(1)})\cr Number of patches.}

\item{patch_size}{(\code{numeric(1)})\cr Patch size.}

\item{init_cohort}{(\code{CohortMat})\cr Initial cohort matrix of class \code{CohortMat}, created by \link{CohortMat}.}

\item{batchsize}{(\code{integer(1)})\cr Batch size, model will be trained in random batch sizes of the data to preserve memory and improve convergence.}

\item{device}{(\code{character(1)})\cr Should the model be fitted on the CPU or the GPU (Graphic card). Support is only for NVIDIA GPUs available.}

\item{debug}{(\code{logical(1)})\cr Debug modus or not. If \code{TRUE}, individual tree states are stored.}

\item{...}{\cr Not used.}
}
\description{
Predict from a FINN model
}
\details{
Simulate from a (fitted) FINN model. This is an S3 method for the
\link[stats:predict]{stats::predict} generic, so it is dispatched as \code{predict(model, ...)}.
}
