% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINN.R
\name{regeneration}
\alias{regeneration}
\title{Calculate the regeneration of forest patches BA_stemsed on the input parameters.}
\usage{
regeneration(species, parReg, pred, light)
}
\arguments{
\item{species}{torch.Tensor species information.}

\item{parReg}{torch.Tensor Regeneration parameters.}

\item{pred}{torch.Tensor Prediction values.}

\item{light}{torch.Tensor Available light variable for calculation.}
}
\value{
torch.Tensor Regeneration values for forest patches.
}
\description{
This function calculates the regeneration of forest patches BA_stemsed on species information, regeneration parameters, prediction values, and available light.
}
