% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINN.R
\name{regeneration}
\alias{regeneration}
\title{Calculate the regeneration of forest patches BA_stemsed on the input parameters.}
\usage{
regeneration(species, parReg, pred, light, patch_size_ha, debug = F)
}
\arguments{
\item{species}{torch.Tensor species information.}

\item{parReg}{torch.Tensor Regeneration parameters. 0 <= parReg <= 1
This parameter denotes the fraction of light needed for a species to regenerate.
In general low values for high regeneration and high values for low regeneration.}

\item{pred}{torch.Tensor Prediction values.}

\item{light}{torch.Tensor Available light variable for calculation.}
}
\value{
torch.Tensor Regeneration values for forest patches.
}
\description{
This function calculates the regeneration of forest patches BA_stemsed on species information, regeneration parameters, prediction values, and available light.
}
