% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createInputs.R
\name{resolveSiteIDs}
\alias{resolveSiteIDs}
\title{Resolve siteIDs into array dimensions needed for the FINN simulation}
\usage{
resolveSiteIDs(tree_dt, env_dt, obs_dt, createInitCohorts = T)
}
\arguments{
\item{site_raw}{A data.table containing raw site-level data with columns "siteID" and "year".}

\item{tree_raw}{A data.table containing raw tree-level data with columns "siteID", "patchID", and "year".}

\item{selection_priority}{A string indicating the priority for selecting sites. Options are "fixed patches" or "many years". Default is "fixed patches".}

\item{Npatches_fixed}{An integer specifying the number of patches to fix when selection_priority is "fixed patches". Default is 4.}

\item{all}{A logical indicating whether to include all sites and patches, even if they do not match. Default is FALSE.}
}
\value{
A data.table with resolved siteID and patchID indices along with their original IDs and years.
}
\description{
This function processes raw site and tree data to create a mapping of siteIDs and patchIDs into integer indices suitable for array dimensions in the FINN model.
}
