% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortMat.R
\name{rweibull_cohorts}
\alias{rweibull_cohorts}
\title{Generate Cohorts Using Weibull Distribution}
\usage{
rweibull_cohorts(
  trees = NULL,
  dbh_shape = NULL,
  dbh_scale = NULL,
  dbh_class_range = 1,
  siteID = 1,
  patchID = 1,
  species = 1
)
}
\arguments{
\item{trees}{Integer vector specifying the number of trees for each cohort. If a single integer is provided, it will be replicated for each draw.}

\item{dbh_shape}{Numeric vector specifying the shape parameters of the Weibull distribution for DBH for each cohort. If a single numeric value is provided, it will be replicated for each draw.}

\item{dbh_scale}{Numeric vector specifying the scale parameters of the Weibull distribution for DBH for each cohort. If a single numeric value is provided, it will be replicated for each draw.}

\item{dbh_class_range}{Numeric value specifying the range of DBH classes. Default is 1.}

\item{siteID}{Integer vector specifying the site ID for each cohort. If a single integer is provided, it will be replicated for each draw. Default is 1.}

\item{patchID}{Integer vector specifying the patch ID for each cohort. If a single integer is provided, it will be replicated for each draw. Default is 1.}

\item{species}{Integer vector specifying the species ID for each cohort. If a single integer is provided, it will be replicated for each draw. Default is 1.}
}
\value{
A data frame containing the cohort data with columns for site ID, patch ID, cohort ID, species, number of trees, and DBH.
}
\description{
This function generates cohort data for tree populations using the Weibull distribution based on specified tree counts, diameter at breast height (DBH) shape, and scale parameters.
}
\details{
The function generates cohort data by drawing samples from the Weibull distribution for each cohort based on the specified shape and scale parameters. The resulting DBH values are binned into classes, and the cohort data is generated accordingly.
}
\examples{
\dontrun{
obs_df <- rweibull_cohorts(
  trees = c(300, 10),
  dbh_shape = c(3, 3),
  dbh_scale = c(1, 50),
  dbh_class_range = 0.1,
  siteID = c(1, 1),
  patchID = c(1, 1),
  species = c(3, 4)
)
CohortMat$new(obs_df = obs_df)
}
}
