% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finn-class.R
\name{simulateForest}
\alias{simulateForest}
\title{Simulate}
\usage{
simulateForest(
  model,
  env,
  disturbance = NULL,
  patches = 100L,
  patch_size = 0.1,
  init_cohort = NULL,
  batchsize = NULL,
  device = c("cpu", "gpu"),
  debug = FALSE
)
}
\arguments{
\item{model}{(\code{finn_class})\cr Object of class \code{finn_class} created by \link{finn}.}

\item{env}{(\code{data.table|data.frame})\cr Data with environmental covariates must be passed as \code{data.table} or \code{data.frame}.}

\item{disturbance}{(\code{data.table|data.frame})\cr Data with disturbance rates must be passed as \code{data.table} or \code{data.frame}.}

\item{patches}{(\code{integer(1)})\cr Number of patches.}

\item{patch_size}{(\code{numeric(1)})\cr Patch size.}

\item{init_cohort}{(\code{CohortMat})\cr Initial cohort matrix of class \code{CohortMat}, created by \link{CohortMat}.}

\item{batchsize}{(\code{integer(1)})\cr Batch size, model will be trained in random batch sizes of the data to preserve memory and improve convergence.}

\item{device}{(\code{character(1)})\cr Should the model be fitted on the CPU or the GPU (Graphic card). Support is only for NVIDIA GPUs available.}

\item{debug}{(\code{logical(1)})\cr Debug modus or not. If \code{TRUE}, individual tree states are stored.}
}
\value{
A named list of simulation results, with the site- and patch-level
state variables and demographic rates in \verb{$long$site} / \verb{$long$patch}
(long format: \code{siteID}, \code{year}, \code{species}, \code{variable}, \code{value}).
}
\description{
Simulate
}
\details{
Simulate from a fitted FINN model. This is a thin, backwards-compatible
alias for \link{predict.finn_class}; new code should call
\code{predict(model, ...)} directly.
}
