% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINNinterface.R
\name{simulateForest}
\alias{simulateForest}
\title{Simulate Forest Dynamics}
\usage{
simulateForest(
  env,
  disturbance = NULL,
  mortalityProcess = NULL,
  growthProcess = NULL,
  regenerationProcess = NULL,
  competitionProcess = NULL,
  speciesPars_ranges = list(parGrowth = rbind(c(0.01, 0.99), c(1, 4)), parMort =
    rbind(c(0.01, 0.99), c(1, 4)), parReg = c(0.01, 0.99), parHeight = c(0.3, 0.7),
    parGrowthEnv = rbind(c(0, 2), c(-2, 2)), parMortEnv = rbind(c(0, 2), c(-2, 2)),
    parRegEnv = rbind(c(0, 2), c(-2, 2))),
  height = NULL,
  patches = 10L,
  patch_size = 0.1,
  sp = 5L,
  init = NULL,
  device = c("cpu", "gpu"),
  parallel = FALSE,
  NGPU = 1,
  batchsize = NULL
)
}
\arguments{
\item{env}{A data frame containing the environmental variables used in the simulation.}

\item{height}{A function or set of parameters defining the height-growth relationship, if applicable.}

\item{patches}{An integer specifying the number of patches in the simulation. Default is 10.}

\item{patch_size}{A numeric value representing the size of each patch. Default is 0.1.}

\item{sp}{An integer specifying the number of species in the simulation. Default is 5.}

\item{init}{A custom initialization object for the simulation, if available.}

\item{device}{A character string specifying whether to use \code{"cpu"} or \code{"gpu"} for computation. Default is \code{"cpu"}.}

\item{parallel}{A logical value indicating whether to run the simulation in parallel. Default is \code{FALSE}.}

\item{NGPU}{An integer specifying the number of GPUs to use if \code{device = "gpu"}. Default is 1.}

\item{mortality}{A process object or formula defining the mortality process.}

\item{growth}{A process object or formula defining the growth process.}

\item{regeneration}{A process object or formula defining the regeneration process.}

\item{competition}{A function defining the competition process, if applicable.}
}
\value{
A list of predicted values for each patch and species, containing arrays of tree diameters and tree counts.
}
\description{
This function simulates forest dynamics, including mortality, growth, regeneration, and competition,
across multiple patches and species.
}
\examples{
env_data <- data.frame(siteID = 1:5, year = 2001:2005, temp = rnorm(5), precip = rnorm(5))
predictions <- simulateForest(env = env_data, patches = 3, sp = 2)

}
